separating app and index
